# Optimise `FormSettings` load